+++ src/module/datasworn.ts
@@ -5,7 +5,7 @@ import { IOracle, IOracleCategory, ironsworn, IRow } from 'dataforged'
 import { max } from 'lodash'
 import { marked } from 'marked'
 import { IronswornActor } from './actor/actor'
-import { hash } from './dataforged'
+import { cleanDollars, hash, renderLinksInMove } from './dataforged'
 import { IronswornItem } from './item/item.js'
 
 const THEME_IMAGES = {
@@ -100,6 +100,7 @@ const FOE_IMAGES = {
 
 const PACKS = [
   'foundry-ironsworn.ironswornitems',
+  'foundry-ironsworn.ironswornmoves',
   'foundry-ironsworn.ironswornassets',
   'foundry-ironsworn.ironsworndelvethemes',
   'foundry-ironsworn.ironsworndelvedomains',
@@ -142,6 +143,7 @@ export async function importFromDatasworn() {
   await Item.createDocuments(movesToCreate, {
     pack: 'foundry-ironsworn.ironswornitems',
   })
+  await processDataforgedMoves()
 
   // Assets
   const assetsJson = await fetch(
@@ -366,6 +368,28 @@ export async function importFromDatasworn() {
   }
   await RollTable.createDocuments(oraclesToCreate, {
     pack: 'foundry-ironsworn.ironswornoracles',
+    keepId: true,
+  })
+}
+
+async function processDataforgedMoves() {
+  const movesToCreate = [] as (ItemDataConstructorData &
+    Record<string, unknown>)[]
+  for (const category of ironsworn['Move Categories']) {
+    for (const move of category.Moves) {
+      renderLinksInMove(move)
+      const cleanMove = cleanDollars(move)
+      movesToCreate.push({
+        _id: hash(move.$id),
+        type: 'sfmove',
+        name: move['Name'],
+        img: 'icons/dice/d10black.svg',
+        data: cleanMove,
+      })
+    }
+  }
+  await Item.createDocuments(movesToCreate, {
+    pack: 'foundry-ironsworn.ironswornmoves',
     keepId: true,
   })
 }